Revert "migrate to mongodb phase 1:"
This reverts commit 78964d2daf68ccce132426d302a09c0ffd18e277.
